requirements.txt
pyminizip
mega.py